+++ cmd/create/create-topic.go
@@ -6,11 +6,12 @@ import (
 	"github.com/spf13/cobra"
 )
 
-var flags operations.CreateTopicFlags
-
 func newCreateTopicCmd() *cobra.Command {
+
+	var flags operations.CreateTopicFlags
+
 	var cmdCreateTopic = &cobra.Command{
-		Use:   "topic",
+		Use:   "topic TOPIC",
 		Short: "create a topic",
 		Args:  cobra.MinimumNArgs(1),
 		Run: func(cmd *cobra.Command, args []string) {