+++ cmd/create/create.go
@@ -6,9 +6,10 @@ func NewCreateCmd() *cobra.Command {
 
 	var cmdCreate = &cobra.Command{
 		Use:   "create",
-		Short: "create topics",
+		Short: "create topics, consumerGroups",
 	}
 
 	cmdCreate.AddCommand(newCreateTopicCmd())
+	cmdCreate.AddCommand(newCreateConsumerGroupCmd())
 	return cmdCreate
 }