+++ cmd/deletion/delete-topic.go
@@ -9,7 +9,7 @@ import (
 func newDeleteTopicCmd() *cobra.Command {
 
 	var cmdDeleteTopic = &cobra.Command{
-		Use:   "topic",
+		Use:   "topic TOPIC",
 		Short: "delete a topic",
 		Args:  cobra.MinimumNArgs(1),
 		Run: func(cmd *cobra.Command, args []string) {
@@ -17,6 +17,7 @@ func newDeleteTopicCmd() *cobra.Command {
 				output.Fail(err)
 			}
 		},
+		ValidArgsFunction: operations.CompleteTopicNames,
 	}
 
 	return cmdDeleteTopic