+++ cmd/describe/describe-topic.go
@@ -6,9 +6,9 @@ import (
 	"github.com/spf13/cobra"
 )
 
-var flags operations.DescribeTopicFlags
-
 func newDescribeTopicCmd() *cobra.Command {
+
+	var flags operations.DescribeTopicFlags
 
 	var cmdDescribeTopic = &cobra.Command{
 		Use:   "topic TOPIC",
@@ -19,6 +19,7 @@ func newDescribeTopicCmd() *cobra.Command {
 				output.Fail(err)
 			}
 		},
+		ValidArgsFunction: operations.CompleteTopicNames,
 	}
 
 	cmdDescribeTopic.Flags().StringVarP(&flags.OutputFormat, "output", "o", flags.OutputFormat, "output format. One of: json|yaml|wide")